requirements.txt
asgiref==3.8.1
Django==5.2.1
django-redis==5.4.0
djangorestframework==3.16.0
psycopg2==2.9.10
python-dotenv==1.1.0
redis==6.1.0
sqlparse==0.5.3
tzdata==2025.2
